+++ app/common/src/main/java/com/cnu/spg/admin/dto/request/AdminUserRoleModifyRequest.java
@@ -1,11 +1,11 @@
-package com.cnu.spg.admin.request;
+package com.cnu.spg.admin.dto.request;
 
 import lombok.Data;
 
 import java.util.Set;
 
 @Data
-public class RequestModifyUserRole {
+public class AdminUserRoleModifyRequest {
     private String id;
     private String username;
     private Set<String> roles;
